Terminada Categoria
Se integro e implemento la clase SweetAlert para hacer los mensajes mas
amigables
se creo Estado.skin para los dropdown de estados
Se crearon las validaciones correspondientes en el modulo de categorias

+++ SistemaFacturacion/GestionCategorias.aspx.cs
@@ -14,16 +14,18 @@ namespace SistemaFacturacion
         Leer,
         Actualizar,
         Eliminar
-
     }
-
+    
     public partial class GestionCategorias : System.Web.UI.Page
     {
         FACTURACIONEntities db = new FACTURACIONEntities();
         private static CRUD operacion = CRUD.Ninguna;
+        SweetAlert message = new SweetAlert(showCancelButton: false);
+
 
         protected void Page_Load(object sender, EventArgs e)
-        {
+        {           
+            ///TODO: Arreglar los nombre de las columnas del grid;
             cargarGvCategorias();
         }
 
@@ -31,6 +33,7 @@ namespace SistemaFacturacion
         {
             ///TODO: EXtraer un metodo para obtener el id del linkbutton
             operacion = CRUD.Actualizar;
+            btnCrear.Enabled = false;
             btnGuardar.Enabled = true;
             btnEliminar.Enabled = true;
             LinkButton lkb = (LinkButton)sender;
@@ -46,49 +49,58 @@ namespace SistemaFacturacion
         protected void btnGuardar_Click(object sender, EventArgs e)
         {
             GuardarCambiosRealizados();
-
         }
 
         private void GuardarCambiosRealizados()
         {
-            try
-            {
-                CATEGORIA categoria = new CATEGORIA();  
 
-                switch (operacion)
+            if (ValidarCampos())
+            {
+                try
                 {
-                    case CRUD.Ninguna:
-                        break;
-                    case CRUD.Crear:
-                        categoria.descripcion = txtDescripcion.Text;
-                        categoria.estado = ddlEstado.SelectedValue;
-                        db.CATEGORIA.Add(categoria);
-                        break;
-                    case CRUD.Actualizar:
-                        categoria = db.CATEGORIA.Find(Int32.Parse(txtId.Text));
-                        categoria.descripcion = txtDescripcion.Text;
-                        categoria.estado = ddlEstado.SelectedValue;
-                        db.Entry(categoria).State = System.Data.EntityState.Modified;
-                        break;
-                    case CRUD.Eliminar:
-                        categoria = db.CATEGORIA.Find(Int32.Parse(txtId.Text));
-                        db.CATEGORIA.Remove(categoria);
-                        break;
-                    default:
-                        break;
-                }
-                db.SaveChanges();
-                LimpiarCampos();
-                cargarGvCategorias();
-            }
-            catch (Exception)
-            {
+                    CATEGORIA categoria = new CATEGORIA();
 
-                throw;
-            }
-            finally
-            {
-                operacion = CRUD.Ninguna;
+                    switch (operacion)
+                    {
+                        case CRUD.Ninguna:
+                            break;
+                        case CRUD.Crear:
+                            categoria.descripcion = txtDescripcion.Text;
+                            categoria.estado = ddlEstado.SelectedValue;
+                            db.CATEGORIA.Add(categoria);
+                            break;
+                        case CRUD.Actualizar:
+                            categoria = db.CATEGORIA.Find(Int32.Parse(txtId.Text));
+                            categoria.descripcion = txtDescripcion.Text;
+                            categoria.estado = ddlEstado.SelectedValue;
+                            db.Entry(categoria).State = System.Data.EntityState.Modified;
+                            break;
+                        case CRUD.Eliminar:
+                            categoria = db.CATEGORIA.Find(Int32.Parse(txtId.Text));
+                            db.CATEGORIA.Remove(categoria);
+                            break;
+                        default:
+                            break;
+                    }
+                    db.SaveChanges();
+                    LimpiarCampos();
+                    cargarGvCategorias();
+                    message.title = "Operación Realizada.";
+                    message.type = "success";
+
+                }
+                catch (Exception)
+                {
+                    message.title = "Ocurrio algún problema, favor de verificar.";
+                    message.type = "error";
+                    throw;
+                }
+                finally
+                {
+                    operacion = CRUD.Ninguna;
+                }
+
+                this.ShowMessage(message);
             }
         }
 
@@ -108,6 +120,8 @@ namespace SistemaFacturacion
 
         private void cargarGvCategorias()
         {
+            ////gvCategorias.DataSource = (from cat in db.CATEGORIA
+            //                           select new { cat.id, cat.descripcion, cat.estado, totalArticulos = cat.ARTICULOS.Count() }).ToList();
             gvCategorias.DataSource = db.CATEGORIA.ToList();
             gvCategorias.DataBind();
         }
@@ -117,20 +131,34 @@ namespace SistemaFacturacion
         /// </summary>
         private void LimpiarCampos()
         {
+
             operacion = CRUD.Ninguna;
             txtId.Text = String.Empty;
             txtDescripcion.Text = String.Empty;
             btnGuardar.Enabled = false;
             btnEliminar.Enabled = false;
+            btnCrear.Enabled = true;
             ddlEstado.SelectedIndex = 0;
+        }
+
+        private bool ValidarCampos()
+        {
+
+            if (String.IsNullOrEmpty(txtDescripcion.Text))
+            {
+                message.title = "El campo Descripción es obligatorio.";
+                message.type = "warning";
+                this.ShowMessage(message);
+                return false;
+            }
+
+            return true;
         }
 
         protected void btnCancelar_Click(object sender, EventArgs e)
         {
             LimpiarCampos();
         }
-
-
 
     }
 }
